:bug: Allow clearing out site description
Currently, site that has been previously configured with a
description can no longer be cleared out. Even if the terraform
configuration indicates empty or nil description, the change will not
be persisted to NSOT. Additionally, every time a terraform plan is run,
it will always detect an update-in-place for this field.

This change fixes this behavior by removing site description's omitempty
tag. This will allow any site descriptions to be cleared out.